removed minion/ubiquiti setup, cleaned up tooling

+++ setup/init.sh
@@ -1,44 +1,43 @@
 #!/bin/bash
 
+set -e
+set -u
+
 OPENNMS_HOST=localhost
-OPENNMS_USER=admin
-OPENNMS_PASS=admin
 OPENNMS_PORT=8980
 
-echo "#### Start deploy process"
+echo "#### Deploy Demo OpenNMS System"
+echo .
+echo .
+echo "The following process will initialize the OpenNMS demo stack. ATTENTION: All existing data will be lost!"
+echo "If you only want to reload new config please follow the README on https://github.com/opennms-forge/demo.opennms.org"
+echo .
+echo "The Do you want to continue?(yes/no)"
+read input
+if [ "$input" == "yes" ]
+ then
 
-# Clear Shorewall
-# Since the default config doesn't allow connections to port 8980, we need to disable it for now
+ # Initialize a new demo instance
+ echo "#### Initializing new Horizon Stack"
+ docker-compose down -v && docker-compose pull && systemctl restart docker && docker-compose up -d
+ echo -n "#### Waiting for OpenNMS to come online"
+ until $(curl -L --output /dev/null --silent --head --fail http://${OPENNMS_HOST}:${OPENNMS_PORT}); do
+   printf '.'
+   sleep 2
+ done
+ echo " DONE"
 
-echo "#### Disabling Shorewall" 
-shorewall clear
+ # Setup HELM and Dashboards
+ echo "#### Setup Grafana HELM Plugin"
+ ./setup-helm.sh
+ echo "#### Installing Grafana Dashboards"
+ ./setup-dashboards.sh
 
-# Initialize a new demo instance
+ # Start importing PRIS requisitions
+ echo "#### Start importing requisitions"
+ docker exec -it demo-horizon /bin/sh -c 'for i in $(/usr/bin/find /opt/opennms/etc/imports -type f -printf "%f\n"); do /opt/opennms/bin/send-event.pl -p '"'url http://pris:8000/requisitions/'"'${i%.*}'"''"' uei.opennms.org/internal/importer/reloadImport; done'
 
-echo "#### Initializing new Horizon Stack"
-
-docker-compose down -v && docker-compose pull && systemctl restart docker && docker-compose up -d
-
-echo -n "#### Waiting for OpenNMS to come online"
-until $(curl -L --output /dev/null --silent --head --fail http://${OPENNMS_HOST}:8980); do
-    printf '.'
-    sleep 2
-done
-echo " DONE"
-
-# Setup HELM and Dashboards
-echo "#### Setup Grafana HELM Plugin"
-./setup-helm.sh
-echo "#### Installing Grafana Dashboards"
-./setup-dashboards.sh
-
-# Start importing PRIS requisitions
-echo "#### Start importing requisitions"
-./reload_all_requisition.sh
-
-# Initialize users
-
-echo "#### Creating user accounts"
-./create_users.sh
-
-echo "Deploy process done"
+ echo "Deploy process done"
+else
+ echo "Aborted"
+fi
